+++ src/main/java/contest/collectingbox/module/publicdata/domain/PublicDataApiInfoRepository.java
@@ -1,4 +1,4 @@
-package contest.collectingbox.module.publicdata;
+package contest.collectingbox.module.publicdata.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
